requirements.txt
streamlit
pandas
joblib
scikit-learn
google-generativeai